hotfix: 충돌 해결 및 popup 추가
